Create 1_burbot_assemble.yaml
This is the file you'll need to run to do the assembly. It will use the my_transcriptome_samples.tsv file (once you've edited it) to build the transcriptome. All I've changed is the output directory it will use (line 3) and the name of the sample file to look for (line 5). To actually run it do `./run_eelpond 1_burbot_assemble.yaml` from inside your eelpond conda environment